disassemble: moving more logic into separate folder
