+++ strcs/decorator.py
@@ -7,8 +7,8 @@ from attrs import define
 
 from . import errors
 from .args_extractor import ArgsExtractor
-from .creation import instantiate
 from .disassemble.base import Type, TypeCache
+from .disassemble.creation import instantiate
 from .meta import Meta
 from .not_specified import NotSpecified, NotSpecifiedMeta
