+++ strcs/disassemble/type_tree.py
@@ -4,9 +4,9 @@ import typing as tp
 from collections import OrderedDict
 from collections.abc import Iterable, Sequence
 
-from .disassemble.base import Field, Type, TypeCache
-from .memoized_property import memoized_property
-from .standard import union_types
+from ..memoized_property import memoized_property
+from ..standard import union_types
+from .base import Field, Type, TypeCache
 
 
 class HasOrigBases(tp.Protocol):
